# design/openapi.yaml
openapi: 3.1.0
info:
  title: parkmitra API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Short-stay parking against idle private capacity, Hyderabad.

    Written at phase 3, BEFORE any implementation, and binding for the rest of the
    project (programme rule: "API contract written before any implementation. It is
    binding for the rest of the project.").

    Two rules that shape every path below:

    - **Money is integer paise.** Never rupees, never a float. `4000` is ₹40.00.
      The client never sends an amount; the server computes it. See ADR-003.
    - **A booking is addressed by its `reference_code`, never by its id.** Holding the
      code is the authorisation to read or modify that booking. Ids are internal and
      never appear in a URL. See ADR-002.

# No authentication scheme applies anywhere in v1: there are no accounts
# (ADR-002). The three /bookings/{reference_code} operations are authorised by
# possession of a ~50-bit reference code supplied as a path parameter, which
# OpenAPI cannot express as a securityScheme. Declaring `security: []` here is
# the accurate statement — "no scheme applies" — rather than leaving it
# undefined and letting a reader assume it was forgotten.
security: []
servers:
  - url: /api
    description: Same-origin route handlers
...
paths:
...
  /bookings:
    post:
      tags: [bookings]
      operationId: createBooking
      summary: Book a bay for a window
      description: |
        The server recomputes the price from the spot's hourly rate and the window
        length — any `amount` in the request body is ignored.

        Concurrency: the database refuses an overlapping booking via an exclusion
        constraint. A losing race returns **409**, not 500. See ADR-001.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingRequest" }
      responses:
        "201":
          description: |
            Booked. The `reference_code` in this response is the only way to read this
            booking again — it is not recoverable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The bay was taken for an overlapping window
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: bay_unavailable
                message: That bay was just taken for part of your window.
        "422":
          description: Window outside the spot's opening hours, or zero/negative length
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  responses:
    BadRequest:
      description: Malformed or missing parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: |
        Rate limited. 20 requests per minute per IP on the booking lookup.
        The `Retry-After` header contains the number of seconds until the
        current window resets.
      headers:
        Retry-After:
          description: Seconds until the rate-limit window resets.
          schema: { type: integer, minimum: 1, maximum: 60 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: rate_limited
            message: Too many requests. Try again later.
  schemas:
...
    Window:
      type: object
      description: Half-open. `end` is exclusive.
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
    BookingRequest:
      type: object
      required: [bay_id, start, end, driver_phone, vehicle_reg]
      description: |
        Note the absence of an `amount` field. This is deliberate — see ADR-003.
      properties:
        bay_id: { type: integer, format: int64 }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        driver_phone:
          type: string
          example: "+919876543210"
          description: Personal data.
        driver_name:
          type: [string, "null"]
          description: Personal data. Optional.
        vehicle_reg:
          type: string
          example: TS09AB1234
          description: Personal data.
    Booking:
      type: object
      required:
        [reference_code, status, window, amount_paise, spot_name, bay_label]
      properties:
        reference_code:
          type: string
          example: 7K2M9QX4TB
          description: Treat as a secret. This is the authorisation for this booking.
        status:
          type: string
          enum: [pending, confirmed, cancelled, expired]
        window: { $ref: "#/components/schemas/Window" }
        amount_paise: { type: integer, example: 8000 }
        spot_name: { type: string }
        bay_label: { type: string, example: B-12 }
        address_line: { type: string }
        arrived_at: { type: [string, "null"], format: date-time }
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code.
          example: bay_unavailable
        message:
          type: string
          description: |
            Human-readable. Must never contain personal data, SQL, or a stack trace.